Support markdown=1
Still some cleanup to do in the postprocessor. For example we need to
better handle nested `<p>` tags. Also needs tests.

However, in the common case, this works well. All `markdown=1` blocks
are normalized to the same insignificant whitespace by the
preprocessor:

    start-tag-placeholder

    content

    end-tag-placeholder

This results in the start and end tags each in a `<p>` tags by itself.
The postprocessor then correctly removes the wrapping `<p>` tags.
And the content gets wrapped in its own `<p>`.

Note that PHP Markdown Extra follows this behavior:

    <div markdown="1">

    *text*

    </div>

    <div markdown="1">
    *text*
    </div>

    <div markdown="1">*text*</div>

Each results in the same normalized output:

    <div>
        <p><em>text</em></p>
    </div>
    <div>
        <p><em>text</em></p>
    </div>
    <div>
        <p><em>text</em></p>
    </div>