Add require_changelog workflow action
This action checks that an update was made to any file in docs/change_log/
but only if changes were made to files in markdown/. Presumably,
any changes outside of markdown/ do not affect the behavior and do not
require a notation in the change_log.

When a change_log entry is missing, a comment is added to the PR informing
the PR author. If a later commit to the PR includes a change_log entry,
then the comment is hidden. A comment should only be added once per PR.