Add require_changelog workflow action (#1051)
This action checks that an update was made to any file in docs/change_log/
but only if changes were made to files in markdown/. Presumably,
any changes outside of markdown/ do not affect the behavior and do not
require a notation in the change_log.

If the proper permissions are available, a comment is added to the PR informing
the PR author that an update to the changelog is missing. However, any PR's from
forks do not have permission on the pull_request event. Unfortunately, the
pull_request_target event doesn't seem to work at all. However, as the action
only attempts to add a comment on failure, we only get a failure in the correct
conditions, even if the error message is related to auth.