Remove check of title text.
Title text is usually generated from body text in the docs. And we have
multiple occasions where the body text is a code span. Therefore, we do
not want the words in the dict as they should always be in code spans.
Therefore, we can ignore title text. It would be rare to manually define
title text anyway.