TOC extension: Add option to generate nested anchor IDs
This feature can be useful when linking to specific subsections of the
resultant document, as the anchor ID will more be more specific to the
header and the headers above it. Unlike the default anchor ID scheme, these
more specific links are less likely to break as additions are made to the
markdown document.

For example, consider the following markdown:
```md
 # Header A
 ## Header A
 ## Header B
 ### Header A
 # Header B
 ## Header A
```

Without the `nested_anchor_ids` setting, the resultant HTML would be:
```html
<h1 id="header-a">Header A</h1>
<h2 id="header-a_1">Header A</h2>
<h2 id="header-b">Header B</h2>
<h3 id="header-a_2">Header A</h3>
<h1 id="header-b_1">Header B</h1>
<h2 id="header-a_3">Header A</h2>
```

With the `nested_anchor_ids` setting, the resultant HTML would be:
```html
<h1 id="header-a">Header A</h1>
<h2 id="header-a-header-a">Header A</h2>
<h2 id="header-a-header-b">Header B</h2>
<h3 id="header-a-header-b-header-a">Header A</h3>
<h1 id="header-b">Header B</h1>
<h2 id="header-b-header-a">Header A</h2>
```