Add GitHub Releases action to deploy workflow
While we don't really need to upload release files to GitHub (as they
are uploaded to PyPI), creating releases does provide a mechanism to
notify users that a new release is available. As GitHub will automaticaly
create files from the repo when none are provided, we upload the official
distribution files instead. Fixes #1365.