Refactor changelog
New changelog is at `docs/changelog.md` and uses the format defined at
<https://keepachangelog.com/>. It only contins releases in the 3.x series.
The old changelog at `docs/change_log/` remains in its old format as an
archive and only contains older releases. The archive is removed from
the site navigation and is only accessable from links in the new changelog.
It is anticipated that only one major release (and its minor releases) will
ever be in the new changelog at a time. Therefore, with the release of a
future version 4.0, the 3.x changelog would be archived as well.

Update the contributing guide to reflect the changes.

Update the changelog validation workflow to use an action which understands the new format.

Update deploy action to include release notes from changelog in a GitHub release.