Fix handling of bogus comments.
As with most implementations, we now pass through bogus comments (as
defined by the HTML Spec) unaltered except that they are HTML escaped.
This deviates from the reference implementation which completely ignores
them. As the reference implementation seems to not have even contemplated
their existence, it is not being used as a reference in this instance.
Fixes #1425.