Ensure lines after heading in loose list are properly detabbed
This is a weird edge case. Normally, detabbing would be handled by
the `ListIndentProcessor`. However, in this one case, that class's
`get_level` method would need to return a different result than in
any other case. As there is no way to easily determine this specific
case from that class, we make the adjustment directly in the
`HashHeaderProcessor` class. Fixes #1433.