Refactor abbr escaping
A alternate fix to #1444. This does not exclude the use of carrots or square 
brackets in abbreviations. It still excludes backslashse, however. I played
with backslashes and it just doesn't make sense to support them as they 
have special meaning in the Markdown, not because of their use in regular
expressions.